split static.js
